переделал отображение вкладок в movie-info. Добавилработу с сервером
