+++ src/components/films-list/films-list.test.js
@@ -2,37 +2,10 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import FilmsList from './films-list.jsx';
 import {Provider} from "react-redux";
-import configureStore from 'redux-mock-store';
-
-const mockStore = configureStore([]);
-
-const filmsMock = [
-  {
-    title: `Bohemian rhapsody`,
-    src: `bohemian-rhapsody.jpg`,
-    poster: `bohemian-rhapsody.jpg`,
-    titlePoster: `bohemian-rhapsody.jpg`,
-    genre: `Comedy`,
-    releaseDate: `2015`,
-    preview: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  },
-  {
-    title: `Bohemian rhapsody`,
-    src: `bohemian-rhapsody.jpg`,
-    poster: `bohemian-rhapsody.jpg`,
-    titlePoster: `bohemian-rhapsody.jpg`,
-    genre: `Comedy`,
-    releaseDate: `2015`,
-    preview: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  },
-];
+import store, {filmsMock} from '../../mock-service';
 
 it(`SnapshotTest FilmList`, () => {
 
-  const store = mockStore({
-    films: filmsMock,
-    genreToFilter: `All genres`
-  });
 
   const tree = renderer
   .create(
