+++ src/components/main/main.test.js
@@ -8,22 +8,42 @@ const mockStore = configureStore([]);
 
 const filmsMock = [
   {
-    title: `Bohemian rhapsody`,
-    src: `bohemian-rhapsody.jpg`,
-    poster: `bohemian-rhapsody.jpg`,
-    titlePoster: `bohemian-rhapsody.jpg`,
-    genre: `Comedy`,
-    releaseDate: `2015`,
-    preview: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
+    name: `War of the Worlds`,
+    poster: `https://htmlacademy-react-3.appspot.com/wtw/static/film/poster/War_of_the_Worlds.jpg`,
+    previewImage: `https://htmlacademy-react-3.appspot.com/wtw/static/film/preview/war-of-the-worlds.jpg`,
+    backgroundImage: `https://htmlacademy-react-3.appspot.com/wtw/static/film/background/War_of_the_Worlds.jpg`,
+    backgroundColor: `#9B7E61`,
+    description: `As Earth is invaded by alien tripod fighting machines, one family fights for survival.`,
+    rating: 9.3,
+    scoresCount: 386834,
+    director: `Steven Spielberg`,
+    starring: [`Tom Cruise`, `Dakota Fanning`, `Tim Robbins`],
+    runTime: 116,
+    genre: `Adventure`,
+    released: 2005,
+    id: 1,
+    isFavorite: false,
+    videoLink: `http://media.xiph.org/mango/tears_of_steel_1080p.webm`,
+    previewVideoLink: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
   },
   {
-    title: `Bohemian rhapsody`,
-    src: `bohemian-rhapsody.jpg`,
-    poster: `bohemian-rhapsody.jpg`,
-    titlePoster: `bohemian-rhapsody.jpg`,
-    genre: `Comedy`,
-    releaseDate: `2015`,
-    preview: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
+    name: `War of the Worlds`,
+    poster: `https://htmlacademy-react-3.appspot.com/wtw/static/film/poster/War_of_the_Worlds.jpg`,
+    previewImage: `https://htmlacademy-react-3.appspot.com/wtw/static/film/preview/war-of-the-worlds.jpg`,
+    backgroundImage: `https://htmlacademy-react-3.appspot.com/wtw/static/film/background/War_of_the_Worlds.jpg`,
+    backgroundColor: `#9B7E61`,
+    description: `As Earth is invaded by alien tripod fighting machines, one family fights for survival.`,
+    rating: 9.3,
+    scoresCount: 386834,
+    director: `Steven Spielberg`,
+    starring: [`Tom Cruise`, `Dakota Fanning`, `Tim Robbins`],
+    runTime: 116,
+    genre: `Adventure`,
+    released: 2005,
+    id: 1,
+    isFavorite: false,
+    videoLink: `http://media.xiph.org/mango/tears_of_steel_1080p.webm`,
+    previewVideoLink: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
   },
 ];
 
@@ -31,8 +51,8 @@ const filmsMock = [
 it(`SnapshotTest Main`, () => {
 
   const store = mockStore({
-    films: filmsMock,
-    APPLICATION: {genre: 'drama'}
+    DATA: {films: filmsMock},
+    APPLICATION: {genre: `drama`}
   });
 
   const tree = renderer
