+++ src/components/movie-card-details/movie-card-details.jsx
@@ -1,169 +1,54 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Tabs from './../tabs/tabs.jsx';
-
 
 const MovieCardDetails = (props) => {
 
-  const {film, changeActivePage, activeTab} = props;
-  const {name, poster, backgroundImage, runTime, genre, released, starring} = film;
+  const {film} = props;
 
   return (
     <React.Fragment>
-      <section className="movie-card movie-card--full">
-        <div className="movie-card__hero">
-          <div className="movie-card__bg">
-            <img src={backgroundImage} alt={name} />
-          </div>
-          <h1 className="visually-hidden">WTW</h1>
-          <header className="page-header movie-card__head">
-            <div className="logo">
-              <a href="main.html" className="logo__link">
-                <span className="logo__letter logo__letter--1">W</span>
-                <span className="logo__letter logo__letter--2">T</span>
-                <span className="logo__letter logo__letter--3">W</span>
-              </a>
-            </div>
-            <div className="user-block">
-              <div className="user-block__avatar">
-                <img src="img/avatar.jpg" alt="User avatar" width={63} height={63} />
-              </div>
-            </div>
-          </header>
-          <div className="movie-card__wrap">
-            <div className="movie-card__desc">
-              <h2 className="movie-card__title">{name}</h2>
-              <p className="movie-card__meta">
-                <span className="movie-card__genre">{genre}</span>
-                <span className="movie-card__year">{released}</span>
-              </p>
-              <div className="movie-card__buttons">
-                <button className="btn btn--play movie-card__button" type="button">
-                  <svg viewBox="0 0 19 19" width={19} height={19}>
-                    <use xlinkHref="#play-s" />
-                  </svg>
-                  <span>Play</span>
-                </button>
-                <button className="btn btn--list movie-card__button" type="button">
-                  <svg viewBox="0 0 19 20" width={19} height={20}>
-                    <use xlinkHref="#add" />
-                  </svg>
-                  <span>My list</span>
-                </button>
-                <a href="add-review.html" className="btn movie-card__button">Add review</a>
-              </div>
-            </div>
-          </div>
+      <div className="movie-card__text movie-card__row">
+        <div className="movie-card__text-col">
+          <p className="movie-card__details-item">
+            <strong className="movie-card__details-name">Director</strong>
+            <span className="movie-card__details-value">Wes Andreson</span>
+          </p>
+          <p className="movie-card__details-item">
+            <strong className="movie-card__details-name">Starring</strong>
+            <span className="movie-card__details-value">
+              {film.starring.map((it) => {
+                return (<span key={it}>{it} <br /></span>);
+              })
+              }
+              {/* Bill Murray, <br />
+              Edward Norton, <br />
+              Jude Law, <br />
+              Willem Dafoe, <br />
+              Saoirse Ronan, <br />
+              Tony Revoloru, <br />
+              Tilda Swinton, <br />
+              Tom Wilkinson, <br />
+              Owen Wilkinson, <br />
+              Adrien Brody, <br />
+              Ralph Fiennes, <br />
+              Jeff Goldblum */}
+            </span>
+          </p>
         </div>
-        <div className="movie-card__wrap movie-card__translate-top">
-          <div className="movie-card__info">
-            <div className="movie-card__poster movie-card__poster--big">
-              <img src={poster} alt={`${name} poster`} width={218} height={327} />
-            </div>
-            <div className="movie-card__desc">
-              <nav className="movie-nav movie-card__nav">
-                <Tabs
-                  changeActivePage={changeActivePage}
-                  activeTab={activeTab}
-                />
-              </nav>
-              <div className="movie-card__text movie-card__row">
-                <div className="movie-card__text-col">
-                  <p className="movie-card__details-item">
-                    <strong className="movie-card__details-name">Director</strong>
-                    <span className="movie-card__details-value">Wes Andreson</span>
-                  </p>
-                  <p className="movie-card__details-item">
-                    <strong className="movie-card__details-name">Starring</strong>
-                    <span className="movie-card__details-value">
-                      {starring.map((it) => {
-                          return(<span key={it}>{it} <br /></span>)
-                        })
-                      }
-                      {/* Bill Murray, <br />
-                      Edward Norton, <br />
-                      Jude Law, <br />
-                      Willem Dafoe, <br />
-                      Saoirse Ronan, <br />
-                      Tony Revoloru, <br />
-                      Tilda Swinton, <br />
-                      Tom Wilkinson, <br />
-                      Owen Wilkinson, <br />
-                      Adrien Brody, <br />
-                      Ralph Fiennes, <br />
-                      Jeff Goldblum */}
-                    </span>
-                  </p>
-                </div>
-                <div className="movie-card__text-col">
-                  <p className="movie-card__details-item">
-                    <strong className="movie-card__details-name">Run Time</strong>
-                    <span className="movie-card__details-value">{runTime}</span>
-                  </p>
-                  <p className="movie-card__details-item">
-                    <strong className="movie-card__details-name">Genre</strong>
-                    <span className="movie-card__details-value">{genre}</span>
-                  </p>
-                  <p className="movie-card__details-item">
-                    <strong className="movie-card__details-name">Released</strong>
-                    <span className="movie-card__details-value">{released}</span>
-                  </p>
-                </div>
-              </div>
-            </div>
-          </div>
+        <div className="movie-card__text-col">
+          <p className="movie-card__details-item">
+            <strong className="movie-card__details-name">Run Time</strong>
+            <span className="movie-card__details-value">{film.runTime}</span>
+          </p>
+          <p className="movie-card__details-item">
+            <strong className="movie-card__details-name">Genre</strong>
+            <span className="movie-card__details-value">{film.genre}</span>
+          </p>
+          <p className="movie-card__details-item">
+            <strong className="movie-card__details-name">Released</strong>
+            <span className="movie-card__details-value">{film.released}</span>
+          </p>
         </div>
-      </section>
-      <div className="page-content">
-        <section className="catalog catalog--like-this">
-          <h2 className="catalog__title">More like this</h2>
-          <div className="catalog__movies-list">
-            <article className="small-movie-card catalog__movies-card">
-              <div className="small-movie-card__image">
-                <img src="img/fantastic-beasts-the-crimes-of-grindelwald.jpg" alt="Fantastic Beasts: The Crimes of Grindelwald" width={280} height={175} />
-              </div>
-              <h3 className="small-movie-card__title">
-                <a className="small-movie-card__link" href="movie-page.html">Fantastic Beasts: The Crimes of Grindelwald</a>
-              </h3>
-            </article>
-            <article className="small-movie-card catalog__movies-card">
-              <div className="small-movie-card__image">
-                <img src="img/bohemian-rhapsody.jpg" alt="Bohemian Rhapsody" width={280} height={175} />
-              </div>
-              <h3 className="small-movie-card__title">
-                <a className="small-movie-card__link" href="movie-page.html">Bohemian Rhapsody</a>
-              </h3>
-            </article>
-            <article className="small-movie-card catalog__movies-card">
-              <div className="small-movie-card__image">
-                <img src="img/macbeth.jpg" alt="Macbeth" width={280} height={175} />
-              </div>
-              <h3 className="small-movie-card__title">
-                <a className="small-movie-card__link" href="movie-page.html">Macbeth</a>
-              </h3>
-            </article>
-            <article className="small-movie-card catalog__movies-card">
-              <div className="small-movie-card__image">
-                <img src="img/aviator.jpg" alt="Aviator" width={280} height={175} />
-              </div>
-              <h3 className="small-movie-card__title">
-                <a className="small-movie-card__link" href="movie-page.html">Aviator</a>
-              </h3>
-            </article>
-          </div>
-        </section>
-        <footer className="page-footer">
-          <div className="logo">
-            <a href="main.html" className="logo__link logo__link--light">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
-          <div className="copyright">
-            <p>© 2019 What to watch Ltd.</p>
-          </div>
-        </footer>
       </div>
     </React.Fragment>
   );
@@ -171,25 +56,11 @@ const MovieCardDetails = (props) => {
 
 MovieCardDetails.propTypes = {
   film: PropTypes.shape({
-    name: PropTypes.string.isRequired,
-    poster: PropTypes.string.isRequired,
-    previewImage: PropTypes.string,
-    backgroundImage: PropTypes.string,
-    backgroundColor: PropTypes.string,
-    decription: PropTypes.string,
-    rating: PropTypes.number,
-    scoresCount: PropTypes.number,
-    director: PropTypes.string,
     starring: PropTypes.array,
     genre: PropTypes.string,
     released: PropTypes.number,
-    id: PropTypes.number,
-    isFavorite: PropTypes.bool,
-    videoLink: PropTypes.string,
-    previewVideoLink: PropTypes.string
+    runTime: PropTypes.number,
   }),
-  changeActivePage: PropTypes.func,
-  activeTab: PropTypes.string
 };
 
 export default MovieCardDetails;
