+++ src/components/tabs/tabs.e2e.test.js
@@ -18,7 +18,7 @@ const mockEvent = {
 it(`Should tabs be click`, () => {
 
   const store = mockStore({
-    APPLICATION: {activePage: 'overview'},
+    APPLICATION: {activePage: `overview`},
   });
 
   const changeActivePage = jest.fn();
