+++ src/reducer/reducer.js
@@ -1,5 +1,5 @@
 import {combineReducers} from 'redux';
-import {reducer as data} from './data/data.js'
+import {reducer as data} from './data/data.js';
 import {reducer as application} from './application/application.js';
 import NameSpace from "./name-space.js";
